requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# Auction_Project/settings.py: 7
# Auction_Project/urls.py: 1,2
# Auction_Project/wsgi.py: 12
# manage.py: 14
# netstore/admin.py: 1
# netstore/apps.py: 1
# netstore/forms.py: 1,2,3,4,6,9
# netstore/migrations/0001_initial.py: 7,8,9
# netstore/migrations/0002_auto_20171012_2143.py: 6,7
# netstore/migrations/0003_auto_20171013_1943.py: 6,7
# netstore/migrations/0004_auto_20171017_1532.py: 6,7
# netstore/migrations/0005_auto_20171018_2012.py: 6,7
# netstore/migrations/0006_auto_20171021_2132.py: 6,7
# netstore/migrations/0007_auto_20171021_2259.py: 6,7
# netstore/migrations/0008_auto_20171021_2303.py: 6,7
# netstore/migrations/0009_auto_20171022_0125.py: 6,7
# netstore/migrations/0010_auto_20171022_0154.py: 6,7
# netstore/migrations/0011_auto_20171022_0334.py: 6,7
# netstore/models.py: 1,2,3
# netstore/tests.py: 1,2,5,6,8
# netstore/views.py: 8,9,10,11,20
Django == 1.11.5

# netstore/views.py: 18
django_filter == 1.0.4

# Auction_Project/urls.py: 3
# netstore/serializers.py: 1
# netstore/views.py: 15,16,17,19
djangorestframework == 3.7.0

# netstore/views.py: 21
schedule == 0.4.3
